Improved Type Checker
The Type Checker is now passing all but three test cases.  This is a
pretty significant improvement over the previous status.  However, there
are a number of issues yet to be resolved:

1) isSubtype().  This method is completely broken and will need
significant changes in the future.

2) Simplification.  Currently, types are not simplified very
effectively.  This needs to be adjusted by introducing a concept of
simplification.

3) Currently source-level expressions have a spare operand for holding
the return type of a given expression.  I think this should be replaced
with a different mechanism.