Updated Expression Form
Have updated the form of expressions so that they no longer include a
"type" operand.  This can be implemented in another way which is
preferable since it somehow was a little ugly.  The only expression form
which retains a type member is the function invocation which is
necessary for determining the proper signature of the function being
invoked.

The other thing implemented in this commit is changing the method
SyntacticItem.numberOfOperands() to be just size().  This is just a
convenient simplification.