Added new implementation of isSubtype()
Have added a new implementation of isSubtype() which is potentially more
powerful.  In particular, it works using the "implicit expansion"
approach.  This is a hybrid approach between rewriting and the use of a
coinductive relation.  What I've realised is that rewriting by itself
doesn't seem to be enough to deal with the recursion.  So, you need a
coninductive relation to handle that, but you also need to simply!