Update Invalid Test Cases.
The invalid test cases were still using the old syntax.  This is now
updated so that they are all valid test cases.  That said, they are all
currently failing because they either pass when they shouldn't (i.e.
because no verification) or because the type checker throws an internal
failure (e.g. because it doesn't support record types yet).