Added tests for array/record subtype checking
Have taken the list of automatically generated test cases used for
subtype testing in WyIL and converted them to run here.  The immediate
advantage is that all *array* tests now pass (which isn't true for the
WyIL subtype operator); however, there are lots of problems with subtype
testing in the presence of records.