Added support for intersecting records
The subtype operator previously failed in the presence of record types.
This is now resolved.  However, there remains a problem which needs to
be resolved: that of records with fields of union type (e.g. {int|null
f}).  Such terms are not "atomic" as they can be constructed in other
ways (i.e. {int|null f} is equivalent to {int f}|{null f}).  We need to
expand such terms in order to obtain a complete algorithm.