Refactored SyntacticItem to contain clone()
All implementations of WyalFile SyntacticItem now employ the clone()
method.  This returns a recursively constructed clone which is detached
from any given heap.  However, at this stage, the returned object does
not reflect the underlying aliasing structure.  This is, in fact, a
problem since we need to retain the aliasing structure around variable
declarations.