Implementation of ArithmeticSimplification
This provides a preliminary and not complete implementation of
ArithmeticSimplification.  There are at least two outstanding issues:

1) Sorting of elements currently assumes tree nodes are allocated.
Unfortunately, for freshly created nodes this is not true.  The planned
solution is to allow nodes to "reconfigure" themselves upon allocation.
This requires an allocation hook to be added to AbstractSyntacticItem.

2) Equality's assume integer operands (which is not always the case).
The key problem here is how to determine the type of operands.  One
solution is to introduce a special "integer equality" operator.  Since
the TypeChecker throws away all other type information, it's not clear
what else to do.  Furthermore, there is a question as to whether or not
other parts of the system will require such information (and, if so, how
to get that).

At this stage, some tests are now passing, though not very many.