Fix for sorting of SyntacticItems
Various syntactic items need to ensure their operands are in sorted
order.  Examples include the conjunction and disjunction items.
Likewise, the polynomial item.  To facility this, I have reverted to a
relatively simple compareTo() operator which operates correctly over
trees (though will not handle cycles).  This is perhaps not super
efficient, though likely it's not a big issue in practice.