InequalityClosure is partially working
The inequality closure rewrite is now starting to work, though there
remain several issues to be resolved.  In particular, there are lots of
limitations with the closure rules themselves:

1) they currently assume inequalities are in a special normal form we
terms on either side are positive.
(which they certainly are not guaranteed to be at this stage)

2) they currently assume polynomials are in a special normal form (which
they are not guaranteed to be under substitution)

3) the rules do not distinguish strict from non-strict inequalities.