Support getType() and further simplification
This adds the getReturnType() function to Expr, which returns the type
that a given expression will evaulate to.  This doesn't yet support
multiple returns, but should do in the future.

This also refactors Formula to exist outside of WyalFile.  And, adds
further simplifications which result in all simple tests now passing.

The hunt is on!!