Support function pre/post conditions.
At the moment, function pre and post conditions only arise because of
type invariants.  But, these are now supported properly by lazily
instantiating them as terms become "known truths".  This is a good point
to do this, since we know they are definitely being considered as "true"
in some sense.

An interesting question remains as to how one should deal with
undefinedness.  At the moment, undefinedness is treated as not existing.
This means that when saying "6 / x == 2", this implicitly implies that x
!= 0.  Therefore, "(x == 0) && (6/x == 2)" reduces to false.