Finished basic (ropey) version
This is the first version to be considered in some sense "working".  Of
course, it's not working that well yet.  But, it is passing a pretty
good number of test cases.

This version provides a starting point to further develop the core
algorithms and procedures, and also to begin the important process of
integrating it into the Whiley compiler itself.  This latter point is
perhaps the next major stage, to basically see where we end up.