Attempt to resolve problem of Macro expansion
A problem with macro expansion at this point is that it does not clone
variable declarations.  This means, for example, that when expanding the
same macro in different positions, you end up potentially with variables
of the same name shared between the two expansions.  This is
particularly problematic if one of the macros was negated and gives rise
to some existentially declared variables, as these can then cause
substitutions into a universally quantified formula.

The changes made in an effort to mitigate this are:

* Updates expandMacroBody() to clone any VariableDeclarations within the
  macro itself.  This is necessary to ensure quantified variables are
  distinct.  At the same time, it must carefully avoid cloning
  VariableDeclarations corresponding to parameters.  This ensures they
  can then be substituted.

* Updates allocate to ensure it preserves structure by using a mapping
  from old items to newly allocated items.  This turns out to be
  necessary as some items are cloned during the process and, hence,
  could not be later encountered during the process in a way that would
  preserve structure.

* Update the mechanism for determining equivalence in
  StructuralEquivalentHeap.  This now relies on the "equals()" method
  for equivalence.  This is important to avoid collapsing otherwise
  identicial VariableDeclarations together.

At this point, the essential problem is that many tests which were
previously passing are now failing.  The exact reason for this remains
unclear, though it is likely due to some bug when implementing the
above.