Refactoring theorem prover loop
This is beginning the process of refactoring the theorem prover loop.
The purpose of doing this is to allow for analysis of a completed branch
(i.e. one which reaches a contradiction).  The purpose of such analysis
is to discard any open splits which are not needed for the contradiction
itself.  This should (in theory) dramatically reduce the overall search
space, as well as provide shorter and more concise proofs.