Add support for disjunct bypass
The so-called "disjunct bypass" optimisation is a potentially very
powerful mechanism for reducing the state space.  This is now
implemented and operational.  However, it's not making heaps of
difference to the overall performance, at least as far as I can tell.

Need to do further investigation into the effect of this optimisation.
In particular, there are further optimisations than can be applied.
And, the dependency cone calculation may be a little coarse.