Fix for the handling of function/macro invocations
The way in which function invocations were handled within the automated
theorem prover was not ideal.  In particular, for the special case of
functions returning boolean values, it was treating them like macros.
This prevent substitution in some places.

To resolve this issue additionally required fixing the TypeChecker to
correctly instantiate a FunctionOrMacro type for a given invocation
(which for some reason it wasn't previously doing).