Fleshed out redesigned AutomatedTheoremProver #51
This fleshes out an initial redesign of the automated theorem prover.
This is intended to implement a number of critical improvements:

1) All rules are now split out of the theorem prover into separate
classes.  This just helps with manageability.  Also, makes it possible
to introduce different kinds of rules at various points.

2) The system now uses an incremental mechanism for deciding when to
apply a give rule.  This means that a rule is not applied more than once
for any given generated proof (well, close to that anyway).

At this stage, a number of initial rules have been ported over from the
old theorem prover.  This is now proving simple things, though more
rules are of course required.