Fixes bug in congruence closure.
This fixes a bug in the rule for congruence closure which was missing
potentially substitutions that had been already added to the set of
"additions".  Instead, the rule now searches backward through the
history looking for existing equalities which can be reapplied.

In addition, I updated the documentation for the rule.