Add support for assignments #53
In the process of trying to get congruence closure to operate as a
single pass, I've added assignments.  Assignments are essentially a way
of registering a "rearranged" equality.

Unfortunately, I also found a fairly severe bug in quantifier
instantiation.  This is now fixed, but it's left a lot of things not
verifying when they should.