Add documentation for TypeTestClosure
This updates the documentation for TypeTestClosure and adds four new test
cases which arose from thinking about it.  Three of those currently fail
because the type checker doesn't perform any flow typing.

In addition, made a minor fix for type test closure which avoids the
TypeSystem.intersect() method.  This method is broken.