Refactored wyal into wyal + wytp
This splits the WyAL package into two packages: WyAL and WyTP.  The
essential idea is that WyAL represents an instance of the theorem
prover framework provided by WyTP.  This allows the framework itself to
be reduced to a smaller core.  It also means that we can then use the
framework without having to go through WyAL.  For example, within the
Whiley compiler we can generate WyTP formulae without having to type
check them via WyAL (as is currently done).