Fixing the Build
Currently the build is broken because the javac compiler is reporting an
ambiguous method invocation, which Eclipse does not thing is ambiguous.
In fact, I don't think it's ambiguous at all.  It maybe a compiler bug!

For now, I have resolved this simply by renaming methods to avoid the
potential ambiguity altogether.  That should do the trick.