Restrict Quantifier Instantiation #67
This restricts quantifier instantiation to only instantiate based on the
binding of proper triggers.  Previously, some things were getting
through that would not be considered triggers in the true sense.  Whilst
they may be occasionally helpful, they should not be necessary and were
leading to a lot of unnecessary instantiations.