Implement Flow-Sensitive Type Checker #62
This completes the flow-sensitive type checker for WyAL.  The design was
loosely based on that from Whiley's FlowTypeChecker and, in particular,
stole the key trick for flow typing through boolean connectives.  That
trick is pretty neat, even if I do say so myself.

In addition, the type checker now looks for "empty" types and reports
errors accordingly.  An empty type is one which is equivalent to void.