Replace simplify() and construct()
This replaces Formulae.simplify() with a proof rule called
"Simplification".  Likewise, this moves AbstractProofRule.construct()
into the CongruenceClosure proof rule.  This simplifies the core of the
system, though perhaps makes it slightly harder to optimise.

At this stage, there remains a problem related to substituting a
polynomial into another polynomial.  The problem is that subtitution
doesn't "flattern" these nested polynomials.