Removed Expr.Polynomial #68
Unclear whether or not it's a good idea that I remove Expr.Polynomial.
However, it's gone now and I'm left with a hulking mess to sort out.

The advantage of getting rid of Expr.Polynomial is that it means I can
factor out this relatively complex code.  Potentially, it could even go
into its own library.