Rework search mechanism in AutomatedTheoremProver
This is a fairly substantial change to the way that the
AutomatedTheoremProver explores the search space.  The key is the
introduction of a mental model based on "branches" through the space.
Specifically, a branch is determined by a "current" proof state, and a
"head" state.  The current state is an ancestor of the head.  The key is
that the difference between these two states represents the differences
between what has been processed (i.e. current) and what is known (i.e.
head).