Add Support for Name Resolution
This adds support for name resolution in the context of an arbitrary
number of files being compiled.  Specifically, name resolution looks
through one or more source files attempting to resolve a given name.

At this stage, a significant issue is that there is no import statement
defined in WyAL.  Therefore, we need to define one...