Updated javadoc; Add TypeSimplification. #84
The essential issue encountered with issue 84 is that of (a lack of) type
simplification.  This updates the documentation in TypeTestClosure to
reflect this and also adds an interface for performing type
simplification.  It's not completely clear at this stage what level of
simplification is going to be required to resolve the issue though.