Add initial interpreter
The interpreter is responsible for interpreting an assertion based on
a given "domain" of values.  The domain is an interface and can be
implemented in different ways.  For example, a simple "small world"
domain can iterate values assuming a small world solution.  In contract,
a "random world" domain could iterate values generated at random.