Enable Associating Assertions with SyntacticItems
This implements something of a kludge which aims to allow one to associate an
arbitrary SyntacticItem with an Assertion.  The reason for this is to make it
possible to recover the SyntacticItem from a given WyilFile when an error occurs
so that, instead of throwing an exception, we can associate a SyntacticMarker
with it.